backend/requirements.txt
boto3
requests
pandas
beautifulsoup4
python-dotenv